styling calendar
